Refactor: Admin dashboard architecture refactoring and standardization
