+++ frontend/src/app/admin/_components/StatsCards/StatsCards.module.css
@@ -47,8 +47,13 @@
 }
 
 @keyframes loading {
-  0% { transform: translateX(-100%); }
-  100% { transform: translateX(100%); }
+  0% {
+    transform: translateX(-100%);
+  }
+
+  100% {
+    transform: translateX(100%);
+  }
 }
 
 .skeletonText {
