--- frontend/src/app/admin/community/AdminCommunityPage.module.css
@@ -1,63 +1,0 @@
-.container {
-  width: 100%;
-  
-  /* 여백 변수 정의 */
-  --tabs-search-gap: var(--space-48);
-  --grid-pagination-gap: var(--space-60);
-}
-
-.header {
-  margin-bottom: 24px;
-}
-
-.title {
-  font-size: 24px;
-  font-weight: 700;
-  margin-bottom: 32px;
-}
-
-.tabs {
-  display: flex;
-  gap: 24px;
-  border-bottom: 1px solid #eee;
-  padding-bottom: 8px;
-}
-
-.tabs button {
-  background: none;
-  border: none;
-  font-size: 14px;
-  color: #888;
-  cursor: pointer;
-  padding-bottom: 12px;
-  position: relative;
-}
-
-.tabs .activeTab {
-  color: #111;
-  font-weight: 600;
-}
-
-.tabs .activeTab::after {
-  content: '';
-  position: absolute;
-  bottom: -9px;
-  left: 0;
-  width: 100%;
-  height: 2px;
-  background-color: #111;
-}
-
-.searchBar {
-  margin-top: var(--tabs-search-gap);
-  margin-bottom: var(--space-32);
-  max-width: 400px;
-}
-
-.cardGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(410px, 1fr));
-  gap: var(--space-24);
-  margin-bottom: var(--grid-pagination-gap);
-}
-
