+++ frontend/src/app/admin/community/_components/CommunityList/CommunityList.module.css
@@ -1,0 +1,19 @@
+.cardGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(410px, 1fr));
+  gap: var(--space-24);
+  margin-bottom: var(--space-60);
+}
+
+.skeletonCard {
+  height: 240px;
+  border-radius: 12px;
+  background-color: var(--color-text-gray-100);
+}
+
+.empty {
+  grid-column: 1 / -1;
+  text-align: center;
+  padding: var(--space-40);
+  color: var(--color-text-gray-500);
+}
