+++ frontend/src/app/admin/community/page.module.css
@@ -1,0 +1,26 @@
+.container {
+  width: 100%;
+  
+  /* 여백 변수 정의 */
+  --tabs-search-gap: var(--space-48);
+  --grid-pagination-gap: var(--space-60);
+}
+
+.header {
+  margin-bottom: 24px;
+}
+
+
+.title {
+  font-size: 24px;
+  font-weight: 700;
+  margin-bottom: 32px;
+}
+
+.paginationArea {
+  display: flex;
+  justify-content: center;
+  margin-top: var(--space-32);
+}
+
+
