--- frontend/src/app/admin/events/_components/EventTable.module.css
@@ -1,218 +1,0 @@
-.tableWrapper {
-  background: var(--color-bg);
-  display: flex;
-  flex-direction: column;
-}
-
-/* ── 1) 탭 + 숨김 체크박스 행 ── */
-.tabRow {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: var(--space-48);
-  margin-bottom: var(--space-32);
-}
-
-.tabs {
-  display: flex;
-  gap: var(--space-24);
-}
-
-.tab {
-  padding: 4px 8px 8px 8px; /* 상단 패딩 추가하여 높이감 조정 */
-  font-family: 'Inter', 'Pretendard', sans-serif;
-  font-size: 16px;
-  font-weight: 500;
-  color: var(--color-text-gray-500, #6b7280);
-  background: none;
-  border: none;
-  border-bottom: 2px solid transparent;
-  cursor: pointer;
-  transition: all var(--transition);
-  margin-bottom: -1px;
-}
-
-.activeTab {
-  color: var(--color-text-gray-900, #111827);
-  font-weight: 700;
-  border-bottom-color: var(--color-text-gray-900, #111827);
-}
-
-.checkboxArea {
-  display: flex;
-  align-items: center;
-}
-
-/* ── 2) 검색바 행 ── */
-.filterRow {
-  display: flex;
-  align-items: center;
-  margin-bottom: var(--space-32);
-}
-
-.searchArea {
-  width: 410px;
-  flex-shrink: 0;
-}
-
-/* ── 3) 카테고리 칩 행 ── */
-.categoryRow {
-  display: flex;
-  align-items: center;
-  margin-bottom: var(--space-32);
-}
-
-.categoryChips {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--space-12);
-}
-
-.chip {
-  display: flex;
-  padding: 8px 16px;
-  justify-content: center;
-  align-items: center;
-  border-radius: 100px;
-  background: var(--color-text-gray-100, #f3f4f6);
-  border: none;
-  color: #4b5563; /* 정확한 Gray-600 명도 적용 */
-  font-family: 'Inter', sans-serif;
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.chip:hover {
-  background: var(--color-text-gray-200, #e5e7eb);
-}
-
-.activeChip {
-  background: var(--color-text-gray-700, #374151) !important;
-  color: var(--color-white, #ffffff) !important;
-}
-
-/* ── 3) 테이블 ── */
-.container {
-  width: 100%;
-  border-top: 1px solid var(--color-text-gray-300);
-}
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.table thead tr {
-  border-bottom: 1px solid var(--color-text-gray-300);
-}
-
-.table th {
-  padding: var(--space-16) var(--space-24);
-  color: var(--color-text-gray-700);
-  font: var(--text-body-medium);
-  text-align: left;
-}
-
-.table tbody tr {
-  border-bottom: 1px solid var(--color-text-gray-100);
-  transition: background var(--transition);
-}
-
-.table tbody tr:hover {
-  background: var(--color-text-gray-50);
-}
-
-.table td {
-  padding: var(--space-16) var(--space-24);
-  font: var(--text-body-medium);
-  color: var(--color-text-gray-900);
-  vertical-align: middle;
-}
-
-/* 피그마 비율에 맞춘 컬럼 설정 */
-.colTitle { width: auto; max-width: 400px; }
-.colAttendees { width: 100px; text-align: center; }
-.colDate { width: 180px; }
-.colStatus { width: 120px; text-align: center; }
-.colVisibility { width: 100px; text-align: center; }
-.colAction { width: 120px; text-align: right; }
-
-.clickableRow {
-  cursor: pointer;
-  transition: background var(--transition);
-}
-
-.clickableRow:hover {
-  background: var(--color-text-gray-50);
-}
-
-.eventTitleText {
-  font: var(--text-body-bold);
-  color: var(--color-text-gray-900, #111827);
-  display: block;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-/* 액션 영역 */
-.actionGroup {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: var(--space-12);
-}
-
-.deleteBtn {
-  border: none;
-  background: transparent;
-  padding: var(--space-4);
-  cursor: pointer;
-  color: var(--color-error);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.deleteBtn:hover {
-  color: var(--color-error); /* 호버 시에도 배경 없이 컬러만 유지하거나 약간의 투명도 조정 */
-  opacity: 0.7;
-}
-
-.moreWrapper {
-  position: relative;
-}
-
-.iconButton {
-  border: none;
-  background: transparent;
-  padding: var(--space-4);
-  cursor: pointer;
-  color: var(--color-text-gray-500);
-  display: flex;
-  align-items: center;
-}
-
-.popover {
-  position: absolute;
-  right: 0;
-  top: 100%;
-  margin-top: 4px;
-  z-index: 100;
-}
-
-/* 상태 메시지 */
-.loadingTd, .emptyTd {
-  text-align: center;
-  padding: var(--space-40);
-  color: var(--color-text-gray-500);
-}
-
-/* 페이지네이션 */
-.paginationArea {
-  display: flex;
-  justify-content: center;
-  margin-top: var(--space-32);
-}
